Added timeout argument
My downloader kept timing out, whether due to my internet connection or what, I do not know. I simply added the timeout argument and this appears to have resolved the issue. Maybe a command line argument for this would be good? I am not using a proxy by the way (which seems to be the general cause for URLOpen issues)